Move wait_for util to root level test/utils
