+++ framework/test/mocha/functional/http/get/accounts/accounts.js
@@ -23,7 +23,7 @@ const BigNum = require('@liskhq/bignum');
 const accountFixtures = require('../../../../fixtures/accounts');
 const SwaggerEndpoint = require('../../../../common/swagger_spec');
 const randomUtil = require('../../../../../utils/random');
-const waitFor = require('../../../../common/utils/wait_for');
+const waitFor = require('../../../../../utils/wait_for');
 const apiHelpers = require('../../../../common/helpers/api');
 const {
 	getNetworkIdentifier,
