+++ framework/test/mocha/functional/http/get/dapps.js
@@ -19,7 +19,7 @@ const Promise = require('bluebird');
 const { transfer, createDapp } = require('@liskhq/lisk-transactions');
 const accountFixtures = require('../../../fixtures/accounts');
 const randomUtil = require('../../../../utils/random');
-const waitFor = require('../../../common/utils/wait_for');
+const waitFor = require('../../../../utils/wait_for');
 const SwaggerEndpoint = require('../../../common/swagger_spec');
 const apiHelpers = require('../../../common/helpers/api');
 const { getNetworkIdentifier } = require('../../../common/network_identifier');
