+++ framework/test/mocha/functional/http/get/delegates.js
@@ -25,7 +25,7 @@ const BigNum = require('@liskhq/bignum');
 const genesisDelegates = require('../../../data/genesis_delegates.json');
 const accountFixtures = require('../../../fixtures/accounts');
 const randomUtil = require('../../../../utils/random');
-const waitFor = require('../../../common/utils/wait_for');
+const waitFor = require('../../../../utils/wait_for');
 const SwaggerEndpoint = require('../../../common/swagger_spec');
 const apiHelpers = require('../../../common/helpers/api');
 const { Slots } = require('../../../../../src/modules/chain/dpos');
