+++ framework/test/mocha/functional/http/get/votes.js
@@ -25,7 +25,7 @@ const BigNum = require('@liskhq/bignum');
 const accountFixtures = require('../../../fixtures/accounts');
 const randomUtil = require('../../../../utils/random');
 const SwaggerEndpoint = require('../../../common/swagger_spec');
-const waitFor = require('../../../common/utils/wait_for');
+const waitFor = require('../../../../utils/wait_for');
 const apiHelpers = require('../../../common/helpers/api');
 const { getNetworkIdentifier } = require('../../../common/network_identifier');
