+++ framework/test/mocha/functional/http/post/1.second_secret.js
@@ -24,7 +24,7 @@ const phases = require('../../../common/phases');
 const accountFixtures = require('../../../fixtures/accounts');
 const apiHelpers = require('../../../common/helpers/api');
 const randomUtil = require('../../../../utils/random');
-const waitFor = require('../../../common/utils/wait_for');
+const waitFor = require('../../../../utils/wait_for');
 const apiCodes = require('../../../../../src/modules/http_api/api_codes');
 const common = require('./common');
 const { getNetworkIdentifier } = require('../../../common/network_identifier');
